requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

hume==0.4.1
numpy==1.26.2
openai==1.11.1
opencv-python==4.8.1.78
pandas==2.1.4
PyAudio==0.2.12
python-dotenv==1.0.1
streamlit==1.29.0
